examples/create-db-from-backup/pyproject.toml
[project]
name = "create-db-from-backup"
version = "0.1.0"
description = ""
authors = [
    {name = "Ori",email = "oriyudilevich@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
